test(vrt): add tooltip tests
