--- tests/components/tooltip/tooltip-closed.browser.test.js
@@ -1,20 +1,0 @@
-import { describe, it, expect } from "vitest";
-import { render } from "vitest-browser-vue";
-import Component from "./tooltip-closed.js";
-import { variants } from "../defaults.js";
-
-describe("Tooltip Closed Component", () => {
-  describe.for(variants)("%s tooltip", async (variant) => {
-    it("Should match screenshot", async () => {
-      const screen = render(Component, {
-        props: {
-          variant,
-        },
-      });
-
-      await document.fonts.ready;
-
-      await expect(screen.getByTestId("tooltip")).toMatchScreenshot();
-    });
-  });
-});
