--- tests/components/tooltip/tooltip-closed.js
@@ -1,39 +1,0 @@
-import { h } from "vue";
-import BasicComponent from "../component/basic-component.js";
-
-export default {
-  props: {
-    variant: {
-      type: String,
-      default: "",
-    },
-    text: {
-      type: Object,
-      default: [
-        h(BasicComponent, { tag: "button", class: "tooltip-toggle" }),
-        h(BasicComponent, {
-          tag: "span",
-          class: "tooltip-description hide",
-          text: "Tips go here",
-        }),
-      ],
-    },
-    attributes: {
-      type: Object,
-      default: () => ({}),
-    },
-  },
-  setup(props) {
-    props.attributes.class = props.attributes.class || [];
-    props.attributes.class.push("tooltip");
-
-    props.attributes["data-testid"] = props.attributes["data-testid"] || [];
-    props.attributes["data-testid"].push("tooltip");
-
-    return () =>
-      h(BasicComponent, {
-        tag: "div",
-        ...props,
-      });
-  },
-};
